use __int64 instead of DWORD

+++ ThreadIn.h
@@ -24,9 +24,10 @@ public:
 			  }
 		  }) {}
 
-	ThreadIn(DWORD pid, __int64 threadStartAddress) : m_currentProcessId(pid), m_thread([this, threadStartAddress]() { Freeze(threadStartAddress, 0); }) {}
+	ThreadIn(__int64 pid, __int64 threadStartAddress)
+		: m_currentProcessId(pid), m_thread([this, threadStartAddress]() { Freeze(threadStartAddress, 0); }) {}
 
-	ThreadIn(DWORD pid, __int64 threadStartAddress, __int64 forwardAddress)
+	ThreadIn(__int64 pid, __int64 threadStartAddress, __int64 forwardAddress)
 		: m_currentProcessId(pid), m_thread([this, threadStartAddress, forwardAddress]() {
 			  while (true) {
 				  Freeze(threadStartAddress, forwardAddress);
